requirements.txt
django
djangorestframework
django-filter
psycopg2-binary
django-cors-headers
python-decouple
